+++ docs/sync.js
@@ -493,18 +493,11 @@ class Credentials {
   static function(payload) {}
 
   /**
-   * Creates credentials from a user API key.
-   * @param {string} key A string identifying the user by API key.
+   * Creates credentials from an API key.
+   * @param {string} key A string identifying the API key.
    * @return {Credentials} An instance of `Credentials` that can be used in {@linkcode Realm.App.logIn}.
    */
-  static userApiKey(token) {}
-
-  /**
-   * Creates credentials from a server API key.
-   * @param {string} key A string identifying the user by API key.
-   * @return {Credentials} An instance of `Credentials` that can be used in {@linkcode Realm.App.logIn}.
-   */
-  static serverApiKey(token) {}
+  static apiKey(token) {}
 
   /**
    * Creates credentials based on an Apple login.