+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 Based on Realm JS v10.21.1: See changelog below for details on enhancements and fixes introduced between this and the previous pre release (which was based on Realm JS v10.19.5).
 
 ### Breaking changes
+* Removal of deprecated functions, which should be replaced with `Realm.Credentials#apiKey`:
+   * `Realm.Credentials#serverApiKey`
+   * `Realm.Credentials#userApiKey`
 * When no object is found calling `Realm#objectForPrimaryKey`, `null` is returned instead of `undefined`
 * Removed deprecated positional arguments to Email/Password authentication functions
     * The following functions now only accept object arguments: